requirements.txt
fastmcp
requests
pydantic
python-dotenv
